+++ docx4j-core/src/main/java/org/docx4j/jaxb/NamespacePrefixMappings.java
@@ -261,6 +261,9 @@ public class NamespacePrefixMappings implements NamespaceContext, org.docx4j.org
     	if (namespaceUri.equals("http://schemas.microsoft.com/office/word/2024/wordml/sdtformatlock")) {
     		return "w16sdtfl";
     	}
+    	if (namespaceUri.equals("http://schemas.microsoft.com/office/word/2026/wordml/cei")) {
+    		return "cei";  // CR-023
+    	}
 
     	if (namespaceUri.equals("http://schemas.microsoft.com/office/powerpoint/2010/main")) {
     		return "p14";
@@ -718,6 +721,8 @@ public class NamespacePrefixMappings implements NamespaceContext, org.docx4j.org
 		// since 11.5.1
 		if (prefix.equals("w16sdtfl"))
 			return "http://schemas.microsoft.com/office/word/2024/wordml/sdtformatlock";		
+		if (prefix.equals("cei"))
+			return "http://schemas.microsoft.com/office/word/2026/wordml/cei";
 		
 		if (prefix.equals("p14"))
 			return "http://schemas.microsoft.com/office/powerpoint/2010/main";